+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `datasets`: Changed `client.Datasets.Create()` to `client.Datasets.CreateOrUpdate()`.
+
 ## [0.2.1] - 2025-12-05
 
 ### Fixed